Moving Clients.jsx to routes folder
