+++ src/AppRoutes.jsx
@@ -3,7 +3,7 @@ import React from 'react';
 import { Routes, Route, Navigate, Outlet } from 'react-router-dom';
 // Page Imports
 import Home from './routes/Home';
-import UserSection from './routes/UserSection';
+import Clients from './routes/Clients';
 import Inbox from './routes/Inbox';
 import Forms from './routes/Forms';
 
@@ -18,7 +18,7 @@ const AppRoutes = ({ isLoggedIn, loadingActive, loadMessages }) => {
     <Routes>
       <Route exact path="/PASS/" element={isLoggedIn ? <Navigate to={path} replace /> : <Home />} />
       <Route element={<ProtectedRoute isLoggedIn={isLoggedIn} />}>
-        <Route path="/PASS/home" element={<UserSection loadingActive={loadingActive} />} />
+        <Route path="/PASS/clients" element={<Clients loadingActive={loadingActive} />} />
         <Route path="/PASS/forms" element={<Forms />} />
         <Route path="/PASS/inbox" element={<Inbox loadMessages={loadMessages} />} />
         <Route path="*" element={<Navigate to={restorePath} replace />} />
