+++ src/components/Clients/index.js
@@ -1,7 +1,6 @@
 import AddClient from './AddClient';
 import AddClientModal from './AddClientModal';
-import Clients from './Clients';
 import ClientList from './ClientList';
 import ClientListTable from './ClientListTable';
 
-export { AddClient, AddClientModal, Clients, ClientList, ClientListTable };
+export { AddClient, AddClientModal, ClientList, ClientListTable };
